requirements.txt
playwright
google-generativeai
python-dotenv
pillow
